매거진 모듈, 지도 모듈 네이밍 및 코드 리팩토링

+++ app/src/main/java/com/limefriends/molde/menu_map/camera_manager/MoldeReportCheckImageActivity.java
@@ -2,10 +2,12 @@ package com.limefriends.molde.menu_map.camera_manager;
 
 import android.content.Intent;
 import android.net.Uri;
+import android.support.v7.app.ActionBar;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
+import android.view.MenuItem;
 import android.view.View;
-import android.widget.ImageButton;
+import android.widget.Button;
 import android.widget.ImageView;
 
 import com.limefriends.molde.R;
@@ -16,15 +18,14 @@ import java.io.File;
 import butterknife.BindView;
 import butterknife.ButterKnife;
 
-public class MoldeReportCheckImage extends AppCompatActivity {
+public class MoldeReportCheckImageActivity extends AppCompatActivity {
     @BindView(R.id.report_check_image)
     ImageView report_check_image;
-    @BindView(R.id.report_check_ok_button)
-    ImageButton report_check_ok_button;
-    @BindView(R.id.report_check_cancel_button)
-    ImageButton report_check_cancel_button;
+
+    Button image_done_button;
 
     private final int ADD_IMAGE = 100;
+    private Uri imagePath;
     private int imageSeq = 0;
 
     private String reportContent;
@@ -41,9 +42,15 @@ public class MoldeReportCheckImage extends AppCompatActivity {
         setContentView(R.layout.map_activity_molde_report_check_image);
         ButterKnife.bind(this);
 
+        getSupportActionBar().setDisplayOptions(ActionBar.DISPLAY_SHOW_CUSTOM);
+        getSupportActionBar().setCustomView(R.layout.default_toolbar);
+        getSupportActionBar().setDisplayHomeAsUpEnabled(true);
+        getSupportActionBar().setHomeButtonEnabled(true);
+        getSupportActionBar().setDisplayShowTitleEnabled(false);
+
         Intent cameraIntent = getIntent();
         imageSeq = cameraIntent.getIntExtra("imageSeq", 1);
-        final Uri uri = cameraIntent.getParcelableExtra("imagePath");
+        imagePath = cameraIntent.getParcelableExtra("imagePath");
         reportContent = cameraIntent.getStringExtra("reportContent");
         reportAddress = cameraIntent.getStringExtra("reportAddress");
         reportDetailAddress = cameraIntent.getStringExtra("reportDetailAddress");
@@ -52,16 +59,19 @@ public class MoldeReportCheckImage extends AppCompatActivity {
         reportLat = cameraIntent.getStringExtra("reportLat");
         reportLng = cameraIntent.getStringExtra("reportLng");
 
-        if(uri != null){
-            report_check_image.setImageURI(uri);
+        if(imagePath != null){
+            report_check_image.setImageURI(imagePath);
         }
 
-        report_check_ok_button.setOnClickListener(new View.OnClickListener() {
+        image_done_button = getSupportActionBar().getCustomView().findViewById(R.id.done_button);
+        image_done_button.setVisibility(View.VISIBLE);
+
+        image_done_button.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
                 Intent intent = new Intent();
                 intent.setClass(getApplicationContext(), MoldeReportActivity.class);
-                intent.putExtra("imagePath", uri);
+                intent.putExtra("imagePath", imagePath);
                 intent.putExtra("imageSeq", imageSeq);
                 intent.putExtra("reportContent", reportContent);
                 intent.putExtra("reportAddress", reportAddress);
@@ -75,16 +85,18 @@ public class MoldeReportCheckImage extends AppCompatActivity {
             }
         });
 
-        report_check_cancel_button.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                File imageFile = new File(uri.getPath());
-                imageFile.delete();
-                sendBroadcast(new Intent(Intent.ACTION_MEDIA_SCANNER_SCAN_FILE, Uri.fromFile(imageFile)));
-                finish();
-            }
-        });
+    }
 
+    @Override
+    public boolean onOptionsItemSelected(MenuItem item) {
+        if (item.getItemId() == android.R.id.home) {
+            File imageFile = new File(imagePath.getPath());
+            imageFile.delete();
+            sendBroadcast(new Intent(Intent.ACTION_MEDIA_SCANNER_SCAN_FILE, Uri.fromFile(imageFile)));
+            finish();
+            return true;
+        }
+        return false;
     }
